Add JavaFX to create GUI version of Jane
Use code from the javafx tutorial for duke and applied it to my chatbot
instead. Created GuiUi class to replace the original CLI Ui class and
exposed it to the javafx classes via JaneGui. Tweaks in the Main,
MainWindow and DialogBox classes to make it work with my Jane chatbot
code instead.